Why isn't it using the index on a LIKE?

+++ words.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if ! test -s /usr/share/dict/words
+then
+	echo Please install words dictionary
+	echo /usr/share/dict/words
+	exit 1
+fi
+
 DB=words.sqlite
 
 rm -f $DB
@@ -9,7 +16,7 @@ rm -f $DB
 	while read w
 	do
 		echo "INSERT INTO words VALUES (\"$w\", length(\"$w\"));"
-	done < british-english
+	done < /usr/share/dict/words
 	echo "CREATE INDEX words_index_1 ON words (word);"
 	echo "COMMIT;"
 ) | sqlite3 --batch $DB
